backend/requirements.txt
Django>=5.0
django-ninja>=1.0
django-cors-headers>=4.3
Pillow>=10.0
python-dotenv>=1.0
